+++ src/CameraGigeAravis.h
@@ -271,7 +271,7 @@
             * @param value New fps value.
             * @return Success status to set fps.
             */
-            bool setFPS(int fps);
+            bool setFPS(double fps);
 
             /**
             * Set device's format.
