+++ src/CameraV4l2.h
@@ -288,7 +288,7 @@
             * @param value New fps value.
             * @return Success status to set fps.
             */
-            bool setFPS(int fps);
+            bool setFPS(double fps);
 
             /**
             * Set device's format.
